added mongo db storing + exception handling + sone unit tests

+++ src/main/java/com/cloudator/interview/services/LocationService.java
@@ -9,7 +9,7 @@ public interface LocationService {
 
     List<Location> getAllLocations() throws IOException;
 
-    long getLocationCode(String locationName) throws IOException;
+    Integer getLocationCode(String locationName) throws IOException;
 
     String concatLocationsId();
 }
